requirements.txt
Jinja2==3.1.2
locust==2.19.1
PyYAML==6.0.1
